delete bookmarks, /bookmarks endpoint
